backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.12.1
argon2-cffi==23.1.0
argon2-cffi-bindings==25.1.0
asgiref==3.11.1
async-timeout==5.0.1
attrs==26.1.0
billiard==4.2.4
boto3==1.34.49
botocore==1.34.162
celery==5.3.6
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
distro==1.9.0
Django==5.0.3
django-cors-headers==4.3.1
django-filter==24.1
django-redis==5.4.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.1
edge-tts==7.2.8
flower==2.0.1
frozenlist==1.8.0
gunicorn==21.2.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
humanize==4.15.0
idna==3.11
inflection==0.5.1
jmespath==1.1.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kombu==5.6.2
lxml==6.0.2
multidict==6.7.1
numpy==2.4.3
openai==1.12.0
packaging==26.0
pandas==3.0.1
pillow==11.1.0
prometheus_client==0.24.1
prompt_toolkit==3.0.52
propcache==0.4.1
psycopg2-binary==2.9.10
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.12.1
pyotp==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-docx==1.1.2
pytz==2026.1.post1
PyYAML==6.0.3
qrcode==8.2
redis==5.0.2
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
s3transfer==0.10.4
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.5
tabulate==0.10.0
tornado==6.5.5
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
whitenoise==6.6.0
yarl==1.23.0
